+++ src/particles/particle_update.hpp
@@ -1,9 +1,11 @@
 #ifndef PARTICLE_UPDATE_HPP_
 #define PARTICLE_UPDATE_HPP_
 
+#include "AMReX_BLProfiler.H"
 #include "AMReX_Extension.H"
 #include "particle_radiation.hpp"
 #include "particle_types.hpp"
+#include "physics_info.hpp"
 
 #if AMREX_SPACEDIM == 3
 
@@ -12,10 +14,17 @@ namespace quokka
 
 // Traits class for specializing particle property update behavior
 template <ParticleType particleType> struct ParticlePropertyUpdateTraits {
-	// Default implementation - does nothing
+	// Default per-particle implementation - does nothing
 	template <typename problem_t, typename ParticleType, int Nout>
 	AMREX_GPU_DEVICE AMREX_FORCE_INLINE static void updateProperties(ParticleType & /*p*/, amrex::Real /*current_time*/,
 									 LuminosityGpuConstTables<Nout> const & /*gpu_tables*/) noexcept
+	{
+		// Default implementation does nothing
+	}
+
+	// Default container-level update - does nothing
+	template <typename problem_t, typename ContainerType>
+	static void updateParticleProperties(ContainerType * /*container*/, amrex::Real /*current_time*/) noexcept
 	{
 		// Default implementation does nothing
 	}
@@ -31,6 +40,38 @@ template <> struct ParticlePropertyUpdateTraits<ParticleType::StochasticStellarP
 	{
 		// Update luminosity using the LuminosityUpdate class
 		LuminosityUpdate::updateLuminosity<problem_t>(p, current_time, gpu_tables);
+	}
+
+	// Container-level update for StochasticStellarPop particles
+	template <typename problem_t, typename ContainerType>
+	static void updateParticleProperties(ContainerType *container, amrex::Real current_time)
+	{
+		const BL_PROFILE("ParticlePropertyUpdateTraits<StochasticStellarPop>::updateParticleProperties()");
+		if (container != nullptr) {
+			constexpr int nGroups = Physics_Traits<problem_t>::nGroups;
+			auto *host_tables_ptr = quokka::g_luminosity_tables_ptr<nGroups>;
+
+			// Only proceed if tables are initialized
+			if (host_tables_ptr != nullptr && host_tables_ptr->is_initialized()) {
+				// Create GPU const tables by value to pass to device
+				auto const gpu_tables = host_tables_ptr->const_tables();
+
+				// Apply the updater to all particles across all levels
+				for (int lev = 0; lev <= container->finestLevel(); ++lev) {
+					for (typename ContainerType::ParIterType pIter(*container, lev); pIter.isValid(); ++pIter) {
+						auto &particles = pIter.GetArrayOfStructs();
+						auto *pData = particles().data();
+						const amrex::Long np = pIter.numParticles();
+
+						amrex::ParallelFor(np, [=] AMREX_GPU_DEVICE(int64_t idx) {
+							auto &p = pData[idx]; // NOLINT
+							ParticlePropertyUpdateTraits<ParticleType::StochasticStellarPop>::template updateProperties<
+							    problem_t, typename ContainerType::ParticleType, nGroups>(p, current_time, gpu_tables);
+						});
+					}
+				}
+			}
+		}
 	}
 };
